CCM NG: Database migration for changing the type of org.libreccm.pagemodels.styles.Dimension#value from float to double
git-svn-id: https://svn.libreccm.org/ccm/ccm_ng@5427 8810af33-2d31-482b-a856-94f89814c4df

Former-commit-id: aa1aef301ebcff2c3cd071b9574607e9f16df1cb

+++ ccm-core/src/test/resources-wildfly-remote-h2-mem/scripts/create_ccm_core_schema.sql
@@ -581,10 +581,10 @@ drop sequence if exists HIBERNATE_SEQUENCE;
     create table CCM_CORE.STYLE_MEDIA_QUERIES (
         MEDIA_QUERY_ID bigint not null,
         MAX_WIDTH_UNIT varchar(255),
-        MAX_WIDTH_VALUE float,
+        MAX_WIDTH_VALUE double,
         MEDIA_TYPE varchar(255),
         MIN_WIDTH_UNIT varchar(255),
-        MIN_WIDTH_VALUE float,
+        MIN_WIDTH_VALUE double,
         primary key (MEDIA_QUERY_ID)
     );
